+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - ktls: fix daemons getting stuck on reading [PR #2606]
 - fix reverse-history-search crash on windows [PR #2639]
+- webui: fix legacy session name setup on new PHP [PR #2693]
 
 ### Documentation
 - Add documentation for the barri image file format [PR #2608]
@@ -2292,4 +2293,5 @@ If you want to migrate from your manually configured disk autochanger to simply 
 [PR #2673]: https://github.com/bareos/bareos/pull/2673
 [PR #2686]: https://github.com/bareos/bareos/pull/2686
 [PR #2692]: https://github.com/bareos/bareos/pull/2692
+[PR #2693]: https://github.com/bareos/bareos/pull/2693
 [unreleased]: https://github.com/bareos/bareos/tree/master